replace float32

+++ backend/ent/comment.go
@@ -23,7 +23,7 @@ type Comment struct {
 	// Content holds the value of the "content" field.
 	Content string `json:"content,omitempty"`
 	// Points holds the value of the "points" field.
-	Points float32 `json:"points,omitempty"`
+	Points float64 `json:"points,omitempty"`
 	// Edges holds the relations/edges for other nodes in the graph.
 	// The values are being populated by the CommentQuery when eager-loading is set.
 	Edges CommentEdges `json:"edges"`
@@ -116,7 +116,7 @@ func (c *Comment) assignValues(columns []string, values []interface{}) error {
 			if value, ok := values[i].(*sql.NullFloat64); !ok {
 				return fmt.Errorf("unexpected type %T for field points", values[i])
 			} else if value.Valid {
-				c.Points = float32(value.Float64)
+				c.Points = value.Float64
 			}
 		}
 	}
